deps: follow jdbc type-model move to jdbc.db.api.type
Mechanical import rewrite: BestFitColumnType/Datatype moved out of the
jdbc.db.dialect.api.type package into the neutral jdbc.db.api.type bundle.
